requirements.txt
docopt-ng>=0.7.0,<0.8.0
lxml>=4.5.0,<5.0.0
svg.path>=4.0.0,<5.0.0